+++ src/internal/CosmicClonesTrigger.hpp
@@ -27,7 +27,6 @@ public:
     void postInit() override;
     void triggerObject(GJBaseGameLayer* layer, const int uniqueID, const gd::vector<int>* remapKeys) override;
     std::vector<std::string> getObjectDetails() override;
-    bool ignoreEditorDuration() override;
 };
 
 class StylePickMenu : public cocos2d::CCMenu {